# swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","summary":"Register a new user","parameters":[],"responses":{"201":{"description":"User successfully registered."},"409":{"description":"Email already exists."}},"tags":["authentication"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","summary":"Login user","parameters":[],"responses":{"200":{"description":"User successfully logged in."},"401":{"description":"Invalid credentials."}},"tags":["authentication"]}},"/api/v1/auth/token":{"post":{"operationId":"AuthController_getTokenWithApiKey","summary":"Get auth token using API key","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiKeyAuthDto"}}}},"responses":{"200":{"description":"Successfully authenticated with API key."},"401":{"description":"Invalid API key."}},"tags":["authentication"]}},"/api/v1/monitoring/health":{"get":{"operationId":"HealthController_check","summary":"System health check","parameters":[],"responses":{"200":{"description":""}},"tags":["monitoring"]}},"/api/v1/flags":{"post":{"operationId":"FeatureFlagController_create","summary":"Create a new feature flag","parameters":[{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"The feature flag has been successfully created."}},"tags":["flags"]},"get":{"operationId":"FeatureFlagController_findAll","summary":"Get all feature flags for a tenant","parameters":[{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Return all feature flags."}},"tags":["flags"]}},"/api/v1/flags/{id}":{"get":{"operationId":"FeatureFlagController_findOne","summary":"Get a feature flag by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Return the feature flag."},"404":{"description":"Feature flag not found."}},"tags":["flags"]},"put":{"operationId":"FeatureFlagController_update","summary":"Update a feature flag","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The feature flag has been successfully updated."},"404":{"description":"Feature flag not found."}},"tags":["flags"]},"delete":{"operationId":"FeatureFlagController_remove","summary":"Delete a feature flag","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The feature flag has been successfully deleted."},"404":{"description":"Feature flag not found."}},"tags":["flags"]}},"/api/v1/flags/key/{key}":{"get":{"operationId":"FeatureFlagController_findByKey","summary":"Get a feature flag by key","parameters":[{"name":"key","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Return the feature flag."},"404":{"description":"Feature flag not found."}},"tags":["flags"]}},"/api/v1/flags/{key}/toggle":{"patch":{"operationId":"FeatureFlagController_toggleFlag","summary":"Toggle a feature flag on/off","parameters":[{"name":"key","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The feature flag has been successfully toggled."},"404":{"description":"Feature flag not found."}},"tags":["flags"]}},"/api/v1/tenants":{"post":{"operationId":"TenantController_create","summary":"Create a new tenant","parameters":[],"responses":{"201":{"description":"The tenant has been successfully created."}},"tags":["tenants"]},"get":{"operationId":"TenantController_findAll","summary":"Get all tenants","parameters":[],"responses":{"200":{"description":"Return all tenants."}},"tags":["tenants"]}},"/api/v1/tenants/{id}":{"get":{"operationId":"TenantController_findOne","summary":"Get a tenant by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Return the tenant."},"404":{"description":"Tenant not found."}},"tags":["tenants"]},"put":{"operationId":"TenantController_update","summary":"Update a tenant","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The tenant has been successfully updated."},"404":{"description":"Tenant not found."}},"tags":["tenants"]},"delete":{"operationId":"TenantController_remove","summary":"Delete a tenant","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The tenant has been successfully deleted."},"404":{"description":"Tenant not found."}},"tags":["tenants"]}},"/api/v1/flags/{featureFlagId}/rules":{"post":{"operationId":"TargetingRuleController_create","summary":"Create a new targeting rule","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"The targeting rule has been successfully created."}},"tags":["targeting-rules"]},"get":{"operationId":"TargetingRuleController_findAll","summary":"Get all targeting rules for a feature flag","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Return all targeting rules."}},"tags":["targeting-rules"]}},"/api/v1/flags/{featureFlagId}/rules/{id}":{"get":{"operationId":"TargetingRuleController_findOne","summary":"Get a specific targeting rule","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Return the targeting rule."},"404":{"description":"Targeting rule not found."}},"tags":["targeting-rules"]},"put":{"operationId":"TargetingRuleController_update","summary":"Update a targeting rule","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The targeting rule has been successfully updated."},"404":{"description":"Targeting rule not found."}},"tags":["targeting-rules"]},"delete":{"operationId":"TargetingRuleController_remove","summary":"Delete a targeting rule","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The targeting rule has been successfully deleted."},"404":{"description":"Targeting rule not found."}},"tags":["targeting-rules"]}},"/api/v1/flags/{featureFlagId}/rules/{id}/enable":{"put":{"operationId":"TargetingRuleController_enable","summary":"Enable a targeting rule","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The targeting rule has been enabled."},"404":{"description":"Targeting rule not found."}},"tags":["targeting-rules"]}},"/api/v1/flags/{featureFlagId}/rules/{id}/disable":{"put":{"operationId":"TargetingRuleController_disable","summary":"Disable a targeting rule","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The targeting rule has been disabled."},"404":{"description":"Targeting rule not found."}},"tags":["targeting-rules"]}},"/api/v1/flags/{featureFlagId}/rules/{id}/percentage":{"put":{"operationId":"TargetingRuleController_updatePercentage","summary":"Update the percentage of a targeting rule","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The percentage has been updated."},"404":{"description":"Targeting rule not found."}},"tags":["targeting-rules"]}},"/api/v1/flags/{featureFlagId}/versions":{"post":{"operationId":"FlagVersionController_create","summary":"Create a new flag version","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"The flag version has been successfully created."}},"tags":["flag-versions"]},"get":{"operationId":"FlagVersionController_findAll","summary":"Get all versions of a feature flag","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Return all flag versions."}},"tags":["flag-versions"]}},"/api/v1/flags/{featureFlagId}/versions/current":{"get":{"operationId":"FlagVersionController_getCurrentVersion","summary":"Get the current version of a feature flag","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Return the current flag version."},"404":{"description":"Feature flag or version not found."}},"tags":["flag-versions"]}},"/api/v1/flags/{featureFlagId}/versions/{version}":{"get":{"operationId":"FlagVersionController_findByVersion","summary":"Get a specific version of a feature flag","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"version","required":true,"in":"path","schema":{"type":"number"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Return the flag version."},"404":{"description":"Version not found."}},"tags":["flag-versions"]}},"/api/v1/flags/{featureFlagId}/versions/{version}/rollback":{"post":{"operationId":"FlagVersionController_rollback","summary":"Rollback to a specific version","parameters":[{"name":"featureFlagId","required":true,"in":"path","schema":{"type":"string"}},{"name":"version","required":true,"in":"path","schema":{"type":"number"}},{"name":"x-tenant-id","in":"header","description":"Tenant ID","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"The flag has been rolled back successfully."},"404":{"description":"Version not found."}},"tags":["flag-versions"]}},"/api/v1/evaluation/{key}":{"post":{"operationId":"EvaluationController_evaluateFlag","summary":"Evaluate a single feature flag","parameters":[{"name":"key","required":true,"in":"path","description":"Feature flag key","schema":{"type":"string"}},{"name":"x-tenant-id","required":true,"in":"header","description":"Tenant ID","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["evaluation"]}},"/api/v1/evaluation/batch":{"post":{"operationId":"EvaluationController_evaluateFlags","summary":"Batch evaluate multiple feature flags","parameters":[{"name":"x-tenant-id","required":true,"in":"header","description":"Tenant ID","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["evaluation"]}},"/api/v1/evaluation/health":{"get":{"operationId":"EvaluationController_healthCheck","summary":"Health check for evaluation API","parameters":[],"responses":{"200":{"description":""}},"tags":["evaluation"]}},"/api/v1/metrics/tenant":{"get":{"operationId":"MetricsController_getTenantMetrics","summary":"Get metrics summary for a tenant","parameters":[{"name":"x-tenant-id","required":true,"in":"header","description":"Tenant ID","schema":{"type":"string"}},{"name":"from","required":false,"in":"query","description":"Start date (ISO format)","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","description":"End date (ISO format)","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["metrics"]}},"/api/v1/metrics/flag":{"get":{"operationId":"MetricsController_getFlagMetrics","summary":"Get metrics for a specific flag","parameters":[{"name":"x-tenant-id","required":true,"in":"header","description":"Tenant ID","schema":{"type":"string"}},{"name":"flagKey","required":true,"in":"query","description":"Feature flag key","schema":{"type":"string"}},{"name":"from","required":false,"in":"query","description":"Start date (ISO format)","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","description":"End date (ISO format)","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["metrics"]}},"/api/v1/metrics/tenant/all":{"get":{"operationId":"MetricsController_getAllTenantsMetrics","summary":"Get metrics summary for all tenants (admin only)","parameters":[{"name":"tenantId","required":true,"in":"query","description":"Tenant ID to get metrics for","schema":{"type":"string"}},{"name":"from","required":false,"in":"query","description":"Start date (ISO format)","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","description":"End date (ISO format)","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["metrics"]}},"/api/v1/admin/tenants/{tenantId}/flags":{"get":{"operationId":"AdminController_getFeatureFlagsByTenant","summary":"Get all feature flags for a specific tenant","parameters":[{"name":"tenantId","required":true,"in":"path","description":"ID of the tenant","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/v1/admin/tenants":{"get":{"operationId":"AdminController_getAllTenants","summary":"Get all tenants","parameters":[],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/v1/admin/flags/copy":{"post":{"operationId":"AdminController_copyFeatureFlag","summary":"Copy a feature flag from one tenant to others","parameters":[],"responses":{"201":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/v1/admin/flags/sync":{"post":{"operationId":"AdminController_syncFeatureFlags","summary":"Sync a feature flag across all tenants","parameters":[{"name":"sourceTenantId","required":true,"in":"query","description":"Source tenant ID for the template flags","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/v1/webhooks/tenant":{"post":{"operationId":"WebhookController_handleTenantWebhook","summary":"Receive tenant webhooks from SaaS platform","parameters":[{"name":"x-signature","required":true,"in":"header","description":"HMAC signature for payload verification","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"Webhook payload with tenant information","content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"201":{"description":""}},"tags":["integrations"]}}},"info":{"title":"Feature Flags API","description":"API for managing feature flags","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ApiKeyAuthDto":{"type":"object","properties":{}}}}}